fixing a busted -R on creds search
I broke this when moving creds to its own file.
