+++ lib/msf/ui/console/command_dispatcher/db.rb
@@ -3,6 +3,7 @@
 require 'rexml/document'
 require 'rex/parser/nmap_xml'
 require 'msf/core/db_export'
+require 'msf/ui/console/command_dispatcher/db_common'
 
 module Msf
 module Ui
@@ -14,7 +15,8 @@ class Db
   require 'tempfile'
 
   include Msf::Ui::Console::CommandDispatcher
-
+  include Msf::Ui::Console::CommandDispatcher::DbCommon
+  
   #
   # The dispatcher's name.
   #
@@ -1712,43 +1714,6 @@ class Db
     print_line "Usage: db_rebuild_cache"
     print_line
     print_line "Purge and rebuild the SQL module cache."
-    print_line
-  end
-
-  #
-  # Set RHOSTS in the +active_module+'s (or global if none) datastore from an array of addresses
-  #
-  # This stores all the addresses to a temporary file and utilizes the
-  # <pre>file:/tmp/filename</pre> syntax to confer the addrs.  +rhosts+
-  # should be an Array.  NOTE: the temporary file is *not* deleted
-  # automatically.
-  #
-  def set_rhosts_from_addrs(rhosts)
-    if rhosts.empty?
-      print_status("The list is empty, cowardly refusing to set RHOSTS")
-      return
-    end
-    if active_module
-      mydatastore = active_module.datastore
-    else
-      # if there is no module in use set the list to the global variable
-      mydatastore = self.framework.datastore
-    end
-
-    if rhosts.length > 5
-      # Lots of hosts makes 'show options' wrap which is difficult to
-      # read, store to a temp file
-      rhosts_file = Rex::Quickfile.new("msf-db-rhosts-")
-      mydatastore['RHOSTS'] = 'file:'+rhosts_file.path
-      # create the output file and assign it to the RHOSTS variable
-      rhosts_file.write(rhosts.join("\n")+"\n")
-      rhosts_file.close
-    else
-      # For short lists, just set it directly
-      mydatastore['RHOSTS'] = rhosts.join(" ")
-    end
-
-    print_line "RHOSTS => #{mydatastore['RHOSTS']}"
     print_line
   end
